Fix issue when int64_t was an unknown type
* added "include <stdint.h>"
* int64_t is replaced with size_t

+++ LuaCppInterface/luatable.h
@@ -5,6 +5,8 @@
 #include <string>
 #include <functional>
 #include <memory>
+#include <stdint.h>
+
 #include "luareference.h"
 
 template<typename SIG>
@@ -73,7 +75,7 @@ public:
 
 	// Get all the keys in the LuaTable
 	void ForAllStringKeys(std::function<void(std::string, LuaType::Value)> stringKeys) const;
-	void ForAllIntegerKeys(std::function<void(int64_t, LuaType::Value)> integerKeys) const;
+	void ForAllIntegerKeys(std::function<void(size_t, LuaType::Value)> integerKeys) const;
 
 };
